Fix problems flagged by Dialyzer.
Kostis Sagonas is a patient guy.  Developers like me must exasperate
him quite a bit: someone who should know better than to forget to use
a tool as valuable as Dialyzer as frequently and as naturally as
breathing.  We in the Erlang community get there, eventually, I hope,
one rap on the knuckles at a time.