Bug 950: refactoring, EUnit tests, new config knob
Refactored the code so that it's possible to export a couple of
useful functions, format_event/3 and limited_fmt/4.  The refactoring
also made Eunit tests for the formatting much easier to write.

See the Edoc at the head of riak_err_app.erl for descriptions of
console_error_type and console_sasl_reports.